refactor: extract tarce logic into a hook
